[Fix][Connector-xugu] Mobile testing methods to JdbcXuguIT
